Improve the error handling for “_getFeedUrl”
The previous approach was too generic and was causing (at least for me)
some misunderstanding. Because even existing my “auto_updater.json” file
I was receiving that generic error message. And the real problem was
occurring while parsing the `res.body` because there was an extra comma
in my JSON file.